test(e2e): expand harness-ui — absorb 13 standalone probes (#317)
Absorb 12 UI-chrome / sidebar / shortcut / popover / palette / settings /
i18n / theme / titlebar / tray / focus / tutorial probes into harness-ui
(plus delete the orphan empty-state-minimal whose body had already been
absorbed). Cuts 13 standalone electron launches; harness-ui now packs
21 cases into one launch (~58s wall, ~1s/case).

Probes absorbed:
  probe-e2e-cwd-popover-recent-unfiltered  -> cwd-popover-recent-unfiltered
  probe-e2e-palette-empty                  -> palette-empty
  probe-e2e-palette-nav                    -> palette-nav
  probe-e2e-settings-open                  -> settings-open
  probe-e2e-search-shortcut-f              -> search-shortcut-f
  probe-e2e-tutorial                       -> tutorial
  probe-e2e-titlebar                       -> titlebar
  probe-e2e-tray                           -> tray
  probe-e2e-focus-orchestration            -> focus-orchestration
  probe-e2e-theme-toggle                   -> theme-toggle
  probe-e2e-language-toggle                -> language-toggle
  probe-e2e-i18n-settings-zh               -> i18n-settings-zh
  probe-e2e-empty-state-minimal            -> deleted (orphan; body already in harness)

Persisted-language hygiene: i18n-settings-zh and language-toggle drive
through the persisted preferences store, which writes localStorage
(`ccsm:preferences`). Their disposers now strip that key in addition to
flipping i18next + main back to en, and the harness setup() does the
same on every boot — so a previous run that left zh persisted no longer
breaks the next launch's English-anchored assertions.
