+++ scripts/harness-ui.mjs
@@ -1,20 +1,33 @@
-// Themed harness — UI cluster, Phase-3.
+// Themed harness — UI cluster.
 //
 // Per docs/e2e/single-harness-brainstorm.md §8 (option B + C). Each case
 // below is the de-duplicated body of one of the per-file probes in
-// scripts/probe-e2e-*.mjs. The original probe files have been left in
-// place with a `// MERGED INTO harness-ui.mjs` marker on line 1 and are
-// excluded from the per-file runner via scripts/run-all-e2e.mjs's
-// MERGED_INTO_HARNESS skip list.
+// scripts/probe-e2e-*.mjs. Absorbed probe files are deleted outright (no
+// breadcrumb files); run-all-e2e.mjs auto-discovers via glob so the runner
+// just stops seeing them.
 //
-// Scope (6 cases — pure UI / store-driven, no real claude.exe required):
-//   - sidebar-align             (probe-e2e-sidebar-align)
-//   - no-sessions-landing       (probe-e2e-no-sessions-landing)
-//   - empty-state-minimal       (probe-e2e-empty-state-minimal)
-//   - a11y-focus-restore        (probe-e2e-a11y-focus-restore)
-//   - shortcut-overlay-opens    (new — UI-1 / #188)
-//   - popover-cross-dismiss     (new — popover-mutex / #221)
-//   - type-scale-snapshot       (new — #225, guards 4-step type token system)
+// Scope:
+//   - sidebar-align                        (probe-e2e-sidebar-align)
+//   - no-sessions-landing                  (probe-e2e-no-sessions-landing)
+//   - empty-state-minimal                  (probe-e2e-empty-state-minimal)
+//   - a11y-focus-restore                   (probe-e2e-a11y-focus-restore)
+//   - shortcut-overlay-opens               (UI-1 / #188)
+//   - popover-cross-dismiss                (popover-mutex / #221)
+//   - type-scale-snapshot                  (#225, 4-step type token system)
+//   - banner-i18n-toggle                   (banner i18n)
+//   - toast-a11y                           (#298 follow-up)
+//   - cwd-popover-recent-unfiltered        (probe-e2e-cwd-popover-recent-unfiltered)
+//   - palette-empty                        (probe-e2e-palette-empty, #117 / #258)
+//   - palette-nav                          (probe-e2e-palette-nav)
+//   - settings-open                        (probe-e2e-settings-open)
+//   - search-shortcut-f                    (probe-e2e-search-shortcut-f)
+//   - tutorial                             (probe-e2e-tutorial)
+//   - titlebar                             (probe-e2e-titlebar)
+//   - tray                                 (probe-e2e-tray)
+//   - focus-orchestration                  (probe-e2e-focus-orchestration)
+//   - theme-toggle                         (probe-e2e-theme-toggle)
+//   - language-toggle                      (probe-e2e-language-toggle)
+//   - i18n-settings-zh                     (probe-e2e-i18n-settings-zh)
 //
 // Related UI probes already absorbed into harness-agent.mjs:
 //   - inputbar-visible, chat-copy, input-placeholder
@@ -676,14 +689,1014 @@ async function caseToastA11y({ win, log }) {
   log('error: role=alert/assertive + glyph + close-only dismiss; info: role=status/polite');
 }
 
+// ---------- cwd-popover-recent-unfiltered ----------
+// Regression probe: opening the StatusBar cwd popover must show the FULL
+// Recent list, regardless of the active session's current cwd. Pre-fix,
+// CwdPopover seeded its query input with the current cwd AND filtered on it.
+async function caseCwdPopoverRecentUnfiltered({ app, win, log, registerDispose }) {
+  const RECENT = ['/proj/foo', '/work/bar', '/code/baz'];
+  const ACTIVE_CWD = RECENT[0];
+
+  // Replace the IPC handler so defaultLoadRecent returns our fixture.
+  await app.evaluate(async ({ ipcMain }, list) => {
+    try { ipcMain.removeHandler('import:recentCwds'); } catch {}
+    ipcMain.handle('import:recentCwds', () => list);
+  }, RECENT);
+  registerDispose(async () => {
+    await app.evaluate(({ ipcMain }) => {
+      try { ipcMain.removeHandler('import:recentCwds'); } catch {}
+    });
+  });
+
+  await win.evaluate((cwd) => {
+    window.__ccsmStore.setState({
+      groups: [{ id: 'g1', name: 'Sessions', collapsed: false, kind: 'normal' }],
+      sessions: [{
+        id: 's1', groupId: 'g1', name: 'Test', state: 'idle', cwd,
+        cwdMissing: false, model: 'claude-sonnet-4-5', agentType: 'claude-code'
+      }],
+      activeId: 's1',
+      tutorialSeen: true
+    });
+  }, ACTIVE_CWD);
+  await win.waitForTimeout(200);
+
+  const trigger = win.locator('[data-cwd-chip]').first();
+  await trigger.waitFor({ state: 'visible', timeout: 10_000 });
+  await trigger.click();
+
+  const dialog = win.getByRole('dialog');
+  await dialog.waitFor({ state: 'visible', timeout: 5_000 });
+
+  await win.waitForFunction(
+    (expected) => {
+      const dlg = document.querySelector('[role="dialog"]');
+      if (!dlg) return false;
+      return dlg.querySelectorAll('[role="option"]').length === expected;
+    },
+    RECENT.length,
+    { timeout: 5_000 }
+  ).catch(async () => {
+    const count = await dialog.locator('[role="option"]').count();
+    const texts = await dialog.locator('[role="option"]').allTextContents();
+    throw new Error(`expected ${RECENT.length} recent options on open, got ${count}: ${JSON.stringify(texts)}`);
+  });
+
+  for (const p of RECENT) {
+    const found = await dialog.locator('[role="option"]').filter({ hasText: p }).count();
+    if (found === 0) throw new Error(`recent entry "${p}" not visible on open`);
+  }
+
+  const input = dialog.getByRole('textbox');
+  const initialValue = await input.inputValue();
+  if (initialValue !== '') throw new Error(`input value should be empty on open, got "${initialValue}"`);
+  const placeholder = await input.getAttribute('placeholder');
+  if (!placeholder || !placeholder.includes('foo')) {
+    throw new Error(`expected placeholder to surface current cwd "${ACTIVE_CWD}", got "${placeholder}"`);
+  }
+
+  await input.fill('bar');
+  await win.waitForFunction(
+    () => {
+      const dlg = document.querySelector('[role="dialog"]');
+      return dlg && dlg.querySelectorAll('[role="option"]').length === 1;
+    },
+    null,
+    { timeout: 3_000 }
+  ).catch(async () => {
+    const count = await dialog.locator('[role="option"]').count();
+    throw new Error(`typing "bar" should filter to 1 option, got ${count}`);
+  });
+
+  await input.fill('');
+  await win.waitForFunction(
+    (expected) => {
+      const dlg = document.querySelector('[role="dialog"]');
+      return dlg && dlg.querySelectorAll('[role="option"]').length === expected;
+    },
+    RECENT.length,
+    { timeout: 3_000 }
+  ).catch(async () => {
+    const count = await dialog.locator('[role="option"]').count();
+    throw new Error(`clearing input should restore all ${RECENT.length} options, got ${count}`);
+  });
+
+  // SCREAMING-strings guard.
+  const recentHeader = dialog.locator('text=/^Recent$/').first();
+  const recentOffender = await recentHeader.evaluate((el) => {
+    return window.getComputedStyle(el).textTransform === 'uppercase' ? el.textContent : null;
+  });
+  if (recentOffender) throw new Error(`"Recent" header is CSS-uppercased — forbidden`);
+
+  log(`open with cwd "${ACTIVE_CWD}" → all ${RECENT.length} visible; "bar" filters to 1; clear restores`);
+}
+
+// ---------- palette-empty ----------
+// CommandPalette empty state (#117): on open, NO results shown until user
+// types. Plus #258 CP3 (no-matches block) and CP4 (kbd hint footer).
+async function casePaletteEmpty({ win, log }) {
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({
+      sessions: [{
+        id: 's-palette-1', name: 'Alpha session', state: 'idle', cwd: '~/alpha',
+        model: 'claude-opus-4', groupId: 'g-default', agentType: 'claude-code'
+      }],
+      groups: [{ id: 'g-default', name: 'Sessions', collapsed: false, kind: 'normal' }],
+      activeId: 's-palette-1',
+      tutorialSeen: true
+    });
+  });
+  await win.waitForTimeout(200);
+
+  await win.evaluate(() => {
+    window.dispatchEvent(
+      new KeyboardEvent('keydown', { key: 'f', code: 'KeyF', ctrlKey: true, bubbles: true })
+    );
+  });
+
+  const searchInput = win.locator('input[placeholder*="Search"]');
+  await searchInput.waitFor({ state: 'visible', timeout: 3000 }).catch(() => {
+    throw new Error('palette did not open via Ctrl+F — search input never appeared');
+  });
+
+  const paletteDialog = win.locator('[role="dialog"]').filter({ has: win.locator('input[placeholder*="Search"]') });
+  const paletteOptions = paletteDialog.locator('[role="option"]');
+  const optionsBeforeType = await paletteOptions.count();
+  if (optionsBeforeType !== 0) {
+    throw new Error(`palette rendered ${optionsBeforeType} option(s) on open; expected 0 until user types`);
+  }
+
+  const hintVisible = await win.getByText(/Type to search/i).isVisible().catch(() => false);
+  if (!hintVisible) throw new Error('empty-state hint "Type to search…" not visible on freshly-opened palette');
+
+  await searchInput.click();
+  await searchInput.fill('alpha');
+  await win.waitForTimeout(150);
+
+  const optionsAfterType = await paletteOptions.count();
+  if (optionsAfterType < 1) throw new Error(`after typing "alpha", expected ≥1 option, got ${optionsAfterType}`);
+  const alphaRowVisible = await paletteOptions.filter({ hasText: 'Alpha session' }).first().isVisible();
+  if (!alphaRowVisible) throw new Error('typing "alpha" did not surface the seeded "Alpha session" row');
+
+  // #258 CP4: kbd hint footer.
+  const kbdHints = paletteDialog.locator('[data-testid="cmd-palette-kbd-hints"]');
+  const kbdHintsVisible = await kbdHints.isVisible().catch(() => false);
+  if (!kbdHintsVisible) throw new Error('kbd hint row [data-testid=cmd-palette-kbd-hints] not visible (#258 CP4)');
+  const hintsText = (await kbdHints.innerText()).replace(/\s+/g, ' ').trim();
+  for (const expected of ['Navigate', 'Select', 'Close']) {
+    if (!hintsText.includes(expected)) throw new Error(`kbd hint row missing label "${expected}" — got: ${hintsText}`);
+  }
+
+  // #258 CP3: no-matches block.
+  await searchInput.fill('zzz-no-such-thing-zzz');
+  await win.waitForTimeout(150);
+  const noMatchesBlock = paletteDialog.locator('[data-testid="cmd-palette-no-matches"]');
+  const noMatchesVisible = await noMatchesBlock.isVisible().catch(() => false);
+  if (!noMatchesVisible) throw new Error('no-matches block not visible after typing nonsense (#258 CP3)');
+  const noMatchesText = await noMatchesBlock.innerText();
+  if (!noMatchesText.includes('No matches')) throw new Error(`no-matches block missing "No matches" copy — got: ${noMatchesText}`);
+  if (!noMatchesText.includes('zzz-no-such-thing-zzz')) throw new Error(`no-matches block did not echo typed query — got: ${noMatchesText}`);
+  const noMatchesSvg = await noMatchesBlock.locator('svg').count();
+  if (noMatchesSvg < 1) throw new Error('no-matches block has no SVG icon (expected SearchX) (#258 CP3)');
+
+  await searchInput.fill('');
+  await searchInput.focus();
+  await searchInput.press('Escape');
+  await win.waitForTimeout(400);
+  const stillOpen = await searchInput.isVisible().catch(() => false);
+  if (stillOpen) throw new Error('palette did not close on Esc');
+
+  log(`empty state OK; "alpha" → ${optionsAfterType} option(s); kbd hints + no-matches block + Esc close`);
+}
+
+// ---------- palette-nav ----------
+// CommandPalette keyboard nav: ↓/↑ moves active row, Enter selects.
+async function casePaletteNav({ win, log }) {
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({
+      sessions: [
+        { id: 's-nav-A', name: 'session alpha', state: 'idle', cwd: '~/a', model: 'claude-opus-4', groupId: 'g-default', agentType: 'claude-code' },
+        { id: 's-nav-B', name: 'session bravo', state: 'idle', cwd: '~/b', model: 'claude-opus-4', groupId: 'g-default', agentType: 'claude-code' }
+      ],
+      groups: [{ id: 'g-default', name: 'Sessions', collapsed: false, kind: 'normal' }],
+      activeId: 's-nav-A',
+      tutorialSeen: true
+    });
+  });
+  await win.waitForTimeout(200);
+
+  await win.evaluate(() => {
+    window.dispatchEvent(
+      new KeyboardEvent('keydown', { key: 'f', code: 'KeyF', ctrlKey: true, bubbles: true })
+    );
+  });
+
+  const searchInput = win.locator('input[placeholder*="Search"]');
+  await searchInput.waitFor({ state: 'visible', timeout: 3000 }).catch(() => {
+    throw new Error('palette did not open via Ctrl+F');
+  });
+
+  await searchInput.click();
+  await searchInput.fill('session');
+  await win.waitForTimeout(150);
+
+  const paletteDialog = win.locator('[role="dialog"]').filter({ has: win.locator('input[placeholder*="Search"]') });
+  const options = paletteDialog.locator('[role="option"]');
+  const optionCount = await options.count();
+  if (optionCount < 2) throw new Error(`expected ≥2 option rows after typing "session", got ${optionCount}`);
+
+  async function activeIndex() {
+    const flags = await options.evaluateAll((els) =>
+      els.map((el) => el.getAttribute('aria-selected') === 'true')
+    );
+    return flags.indexOf(true);
+  }
+
+  let idx = await activeIndex();
+  if (idx !== 0) throw new Error(`initial active index expected 0, got ${idx}`);
+
+  await searchInput.press('ArrowDown');
+  await win.waitForTimeout(80);
+  idx = await activeIndex();
+  if (idx !== 1) throw new Error(`after ArrowDown, expected active=1, got ${idx}`);
+
+  await searchInput.press('ArrowUp');
+  await win.waitForTimeout(80);
+  idx = await activeIndex();
+  if (idx !== 0) throw new Error(`after ArrowUp, expected active=0, got ${idx}`);
+
+  const labels = await options.evaluateAll((els) => els.map((el) => el.textContent?.trim() ?? ''));
+  const bravoIdx = labels.findIndex((l) => l.includes('bravo'));
+  if (bravoIdx < 0) throw new Error(`"session bravo" row not present: ${JSON.stringify(labels)}`);
+  const steps = bravoIdx - (await activeIndex());
+  for (let i = 0; i < steps; i++) await searchInput.press('ArrowDown');
+  await win.waitForTimeout(80);
+
+  await searchInput.press('Enter');
+  await win.waitForTimeout(400);
+
+  const activeId = await win.evaluate(() => window.__ccsmStore.getState().activeId);
+  if (activeId !== 's-nav-B') throw new Error(`Enter on "session bravo" did not select s-nav-B; activeId=${activeId}`);
+  const stillOpen = await searchInput.isVisible().catch(() => false);
+  if (stillOpen) throw new Error('palette did not close after Enter');
+
+  log('Ctrl+F opens; ↓/↑ moves active row; Enter selects s-nav-B');
+}
+
+// ---------- settings-open ----------
+// Settings dialog open/close — three entry points reach ONE dialog.
+async function caseSettingsOpen({ win, log }) {
+  // Need a session for the InputBar to render (entry #2 uses /config in textarea).
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({
+      groups: [{ id: 'g1', name: 'G1', collapsed: false, kind: 'normal' }],
+      sessions: [{ id: 's1', name: 's', state: 'idle', cwd: 'C:/x', model: 'claude-opus-4', groupId: 'g1', agentType: 'claude-code' }],
+      activeId: 's1',
+      messagesBySession: { s1: [] },
+      tutorialSeen: true
+    });
+  });
+  await win.waitForTimeout(200);
+
+  const dialog = win.getByRole('dialog');
+
+  async function expectDialogClosed(label) {
+    await dialog.waitFor({ state: 'hidden', timeout: 1500 }).catch(() => {});
+    if ((await dialog.count()) > 0) throw new Error(`${label}: dialog still in DOM after expected close`);
+  }
+  async function expectSettingsDialogOpen(label) {
+    await dialog.waitFor({ state: 'visible', timeout: 3000 }).catch(() => {
+      throw new Error(`${label}: dialog never became visible`);
+    });
+    const conn = dialog.getByRole('tab', { name: /^connection$/i });
+    await conn.waitFor({ state: 'visible', timeout: 1500 }).catch(() => {
+      throw new Error(`${label}: Settings tabs not visible — wrong dialog opened?`);
+    });
+  }
+  async function pressEscAndExpectClosed(label) {
+    await win.keyboard.press('Escape');
+    await expectDialogClosed(label);
+  }
+
+  // 1. Sidebar Settings button.
+  const sidebarBtn = win.getByRole('button', { name: /^settings$/i }).first();
+  await sidebarBtn.waitFor({ state: 'visible', timeout: 5000 });
+  await sidebarBtn.click();
+  await expectSettingsDialogOpen('sidebar button');
+  await pressEscAndExpectClosed('sidebar button');
+
+  // 2. /config slash command.
+  const textarea = win.locator('textarea').first();
+  await textarea.waitFor({ state: 'visible', timeout: 5000 });
+  await textarea.click();
+  await textarea.fill('/config');
+  await win.waitForTimeout(80);
+  await win.keyboard.press('Escape');
+  await win.waitForTimeout(80);
+  await win.keyboard.press('Enter');
+  await expectSettingsDialogOpen('/config');
+  await pressEscAndExpectClosed('/config');
+
+  // 3. Keyboard shortcut Cmd/Ctrl+,.
+  const accel = process.platform === 'darwin' ? 'Meta' : 'Control';
+  await textarea.fill('');
+  await win.locator('body').click({ position: { x: 5, y: 5 } }).catch(() => {});
+  await win.keyboard.press(`${accel}+,`);
+  await expectSettingsDialogOpen('keyboard shortcut');
+  await pressEscAndExpectClosed('keyboard shortcut');
+
+  log('sidebar / /config / Cmd+, all open Settings; Esc closes');
+}
+
+// ---------- search-shortcut-f ----------
+// Cmd/Ctrl+F opens palette, toggles closed; Cmd/Ctrl+K does NOT open it.
+async function caseSearchShortcutF({ win, log }) {
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({
+      groups: [{ id: 'g1', name: 'G1', collapsed: false, kind: 'normal' }],
+      sessions: [{ id: 's1', name: 's', state: 'idle', cwd: 'C:/x', model: 'claude-opus-4', groupId: 'g1', agentType: 'claude-code' }],
+      activeId: 's1',
+      tutorialSeen: true
+    });
+  });
+  await win.waitForTimeout(200);
+
+  const searchInput = win.locator('input[placeholder*="Search"]');
+  const accel = process.platform === 'darwin' ? 'Meta' : 'Control';
+
+  // 1. Open via Cmd/Ctrl+F.
+  await win.keyboard.press(`${accel}+f`);
+  await searchInput.waitFor({ state: 'visible', timeout: 5000 }).catch(() => {
+    throw new Error('palette did not open via Ctrl+F — search input never appeared');
+  });
+
+  // 2. Toggle closed via the same shortcut.
+  await searchInput.focus();
+  await win.keyboard.press(`${accel}+f`);
+  await searchInput.waitFor({ state: 'hidden', timeout: 1500 }).catch(() => {
+    throw new Error('palette did not close on second Ctrl+F (toggle broken)');
+  });
+
+  // 3. Cmd/Ctrl+K must NOT open it.
+  await win.keyboard.press(`${accel}+k`);
+  await win.waitForTimeout(500);
+  const openedByK = await searchInput.isVisible().catch(() => false);
+  if (openedByK) throw new Error('Ctrl+K opened the palette — the K binding for search should be removed');
+
+  log('Ctrl+F opens; Ctrl+F toggles closed; Ctrl+K does NOT open');
+}
+
+// ---------- tutorial ----------
+// First-run tutorial: shows when sessions=[] && tutorialSeen=false; Step 1/4
+// indicator + Skip button; Next advances; Done flips tutorialSeen.
+async function caseTutorial({ win, log }) {
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({ sessions: [], activeId: undefined, tutorialSeen: false });
+  });
+  await win.waitForTimeout(200);
+
+  const stepCounter = win.locator('text=/Step 1 of 4/i').first();
+  await stepCounter.waitFor({ state: 'visible', timeout: 5000 });
+  await win.locator('text=/A workbench for AI sessions/i').first().waitFor({ state: 'visible', timeout: 3000 });
+
+  // SCREAMING-strings guard (PR #248 Gap #1, task #315).
+  const screaming = await win.evaluate(() => {
+    const root = document.querySelector('[data-testid="tutorial"], main, body');
+    if (!root) return [];
+    const offenders = [];
+    const walker = document.createTreeWalker(root, NodeFilter.SHOW_ELEMENT);
+    let node = walker.currentNode;
+    while (node) {
+      const el = node;
+      if (el && el.textContent && el.children.length === 0) {
+        const txt = el.textContent.trim();
+        if (txt && /[a-zA-Z]/.test(txt)) {
+          const tt = window.getComputedStyle(el).textTransform;
+          if (tt === 'uppercase') offenders.push(`${el.tagName}: ${txt.slice(0, 60)}`);
+        }
+      }
+      node = walker.nextNode();
+    }
+    return offenders;
+  });
+  if (screaming.length > 0) {
+    throw new Error(`tutorial has CSS-uppercased text (forbidden):\n  ${screaming.join('\n  ')}`);
+  }
+
+  const skipBtn = win.getByRole('button', { name: /^Skip$/ });
+  await skipBtn.waitFor({ state: 'visible', timeout: 3000 });
+
+  const nextBtn = win.getByRole('button', { name: /^Next$/ });
+  await nextBtn.click();
+  await win.waitForTimeout(150);
+  await nextBtn.click();
+  await win.waitForTimeout(150);
+  await nextBtn.click();
+  await win.waitForTimeout(200);
+
+  await win.locator('text=/Step 4 of 4/i').first().waitFor({ state: 'visible', timeout: 3000 });
+  await win.locator('text=/Ready when you are/i').first().waitFor({ state: 'visible', timeout: 3000 });
+
+  await win.getByRole('button', { name: /^New Session$/ }).first().waitFor({ state: 'visible', timeout: 3000 });
+  await win.getByRole('button', { name: /^Import Session$/ }).first().waitFor({ state: 'visible', timeout: 3000 });
+
+  await win.getByRole('button', { name: /^Done$/ }).click();
+  await win.waitForTimeout(300);
+
+  const seen = await win.evaluate(() => window.__ccsmStore.getState().tutorialSeen);
+  if (!seen) throw new Error('Done did not set tutorialSeen=true');
+
+  const stillTutorial = await win.locator('text=/Step \\d of 4/i').count();
+  if (stillTutorial > 0) throw new Error('tutorial still rendered after Done');
+
+  log('Step 1→4 advance; Done sets tutorialSeen=true');
+}
+
+// ---------- titlebar ----------
+// No native frame on win/linux; >=2 top drag regions of expected height;
+// window controls inside right pane (not sidebar) on win/linux.
+async function caseTitlebar({ app, win, log }) {
+  await win.waitForFunction(() => !!document.querySelector('main') && !!document.querySelector('aside'), null, { timeout: 10000 });
+
+  const platform = await app.evaluate(() => process.platform);
+
+  const dragRegions = await win.evaluate(() => {
+    const els = Array.from(document.querySelectorAll('[style*="app-region: drag"]'));
+    return els.map((el) => {
+      const r = el.getBoundingClientRect();
+      return { height: Math.round(r.height), top: Math.round(r.top), left: Math.round(r.left), width: Math.round(r.width) };
+    }).filter((r) => r.top === 0);
+  });
+  if (dragRegions.length < 2) {
+    throw new Error(`expected >=2 top drag regions, got ${dragRegions.length}: ${JSON.stringify(dragRegions)}`);
+  }
+  const EXPECTED_STRIP = 32;
+  for (const r of dragRegions) {
+    if (Math.abs(r.height - EXPECTED_STRIP) > 2) {
+      throw new Error(`drag region height expected ~${EXPECTED_STRIP}, got ${r.height}. all=${JSON.stringify(dragRegions)}`);
+    }
+  }
+
+  if (platform !== 'darwin') {
+    for (const name of ['Minimize', 'Close']) {
+      await win.locator(`button[aria-label="${name}"]`).waitFor({ state: 'visible', timeout: 5000 });
+    }
+    await win.locator('button[aria-label="Maximize"], button[aria-label="Restore"]').first().waitFor({ state: 'visible', timeout: 5000 });
+
+    const geometry = await win.evaluate(() => {
+      const sidebar = document.querySelector('aside');
+      const close = Array.from(document.querySelectorAll('button')).find((b) => b.getAttribute('aria-label') === 'Close');
+      if (!sidebar || !close) return null;
+      const s = sidebar.getBoundingClientRect();
+      const c = close.getBoundingClientRect();
+      return { sidebarRight: s.right, closeLeft: c.left, closeRight: c.right, windowWidth: window.innerWidth };
+    });
+    if (!geometry) throw new Error('sidebar or Close button missing');
+    if (geometry.closeLeft < geometry.sidebarRight) {
+      throw new Error(`Close button is not inside the right pane (closeLeft=${geometry.closeLeft} < sidebarRight=${geometry.sidebarRight})`);
+    }
+    if (geometry.windowWidth - geometry.closeRight > 4) {
+      throw new Error(`Close button not flush to window right edge (gap=${geometry.windowWidth - geometry.closeRight})`);
+    }
+  }
+
+  log(`platform=${platform} dragRegions=${dragRegions.length}`);
+}
+
+// ---------- tray ----------
+// Closing the window hides it (does NOT quit) on win32/linux; show() restores.
+async function caseTray({ app, win, log, registerDispose }) {
+  // Belt-and-suspenders: ensure the window is visible after we're done so
+  // subsequent cases can interact with it (the harness window is shared).
+  registerDispose(async () => {
+    await app.evaluate(({ BrowserWindow }) => {
+      const w = BrowserWindow.getAllWindows().find((x) => !x.webContents.getURL().startsWith('devtools://'));
+      try { w?.show(); } catch {}
+    });
+  });
+
+  await app.evaluate(({ BrowserWindow }) => {
+    const w = BrowserWindow.getAllWindows().find((x) => !x.webContents.getURL().startsWith('devtools://'));
+    w?.close();
+  });
+  // Main process fades for 180ms before hide().
+  await new Promise((r) => setTimeout(r, 600));
+
+  const state = await app.evaluate(({ BrowserWindow }) => {
+    const w = BrowserWindow.getAllWindows().find((x) => !x.webContents.getURL().startsWith('devtools://'));
+    return { exists: !!w, visible: w?.isVisible() ?? null };
+  });
+  if (!state.exists) throw new Error('window was destroyed; expected hide-on-close');
+  if (state.visible !== false) throw new Error(`window should be hidden after close; visible=${state.visible}`);
+
+  await app.evaluate(({ BrowserWindow }) => {
+    const w = BrowserWindow.getAllWindows().find((x) => !x.webContents.getURL().startsWith('devtools://'));
+    w?.show();
+  });
+  const after = await app.evaluate(({ BrowserWindow }) => {
+    const w = BrowserWindow.getAllWindows().find((x) => !x.webContents.getURL().startsWith('devtools://'));
+    return w?.isVisible() ?? null;
+  });
+  if (after !== true) throw new Error(`window should be visible after show; visible=${after}`);
+
+  log('hide-on-close=true; restore-via-show=true');
+}
+
+// ---------- focus-orchestration ----------
+// Composer focus contracts: send / switch-session / group-click / modal /
+// preserve-other-input. See probe-e2e-focus-orchestration.mjs for full notes.
+async function caseFocusOrchestration({ win, log }) {
+  await win.evaluate(() => {
+    try {
+      const real = window.ccsm;
+      if (real) {
+        real.agentSend = async () => true;
+        real.agentStart = async () => ({ ok: true, sessionId: 'sdk-1' });
+        real.agentInterrupt = async () => true;
+      }
+    } catch {}
+  });
+
+  await win.evaluate(() => {
+    window.__ccsmStore.setState({
+      groups: [
+        { id: 'g1', name: 'Group One', collapsed: false, kind: 'normal' },
+        { id: 'g2', name: 'Group Two', collapsed: false, kind: 'normal' }
+      ],
+      sessions: [
+        { id: 'sA', name: 'session-a', state: 'idle', cwd: 'C:/x', model: 'claude-opus-4', groupId: 'g1', agentType: 'claude-code' },
+        { id: 'sB', name: 'session-b', state: 'idle', cwd: 'C:/y', model: 'claude-opus-4', groupId: 'g2', agentType: 'claude-code' }
+      ],
+      activeId: 'sA',
+      messagesBySession: { sA: [{ kind: 'user', id: 'u-a', text: 'hi A' }], sB: [] },
+      startedSessions: { sA: true }
+    });
+  });
+  await win.waitForTimeout(300);
+
+  const textarea = win.locator('textarea');
+  await textarea.waitFor({ state: 'visible', timeout: 5000 });
+
+  const activeTag = () => win.evaluate(() => document.activeElement ? document.activeElement.tagName : null);
+  const activeIsTextarea = () => win.evaluate(() => document.activeElement?.tagName === 'TEXTAREA');
+
+  // Contract 1: focus returns (or is acceptable as BUTTON) after Send.
+  await textarea.click();
+  await textarea.fill('hello there');
+  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
+  if (await activeIsTextarea()) throw new Error('precondition: textarea should not be focused right before Send');
+  const sendBtn = win.getByRole('button', { name: /send message/i }).first();
+  await sendBtn.waitFor({ state: 'visible', timeout: 3000 });
+  await sendBtn.click();
+  await win.waitForTimeout(200);
+  const echoLanded = await win.evaluate(() => {
+    const blocks = window.__ccsmStore.getState().messagesBySession['sA'] ?? [];
+    return blocks.some((b) => b.kind === 'user' && b.text === 'hello there');
+  });
+  if (!echoLanded) throw new Error('Send click did not produce a user-echo block — send() never ran');
+  // BUTTON-after-send is currently soft-acceptable; we just log.
+  const tagAfterSend = await activeTag();
+  log(`post-Send activeElement=${tagAfterSend}`);
+
+  // Contract 2: switching session via sidebar focuses target textarea.
+  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
+  const rows = win.locator('aside li[role="option"]');
+  const rowCount = await rows.count();
+  if (rowCount < 2) throw new Error(`expected ≥2 sidebar rows, got ${rowCount}`);
+  await rows.nth(1).click();
+  await win.waitForTimeout(150);
+  if (!(await activeIsTextarea())) {
+    const tag = await activeTag();
+    throw new Error(`after switching session via sidebar click, expected TEXTAREA focus, got ${tag}`);
+  }
+
+  // Contract 3: clicking a group header doesn't drop focus to <body>.
+  await textarea.click();
+  await textarea.fill('drafting…');
+  if (!(await activeIsTextarea())) throw new Error('precondition: textarea should be focused before group click');
+  const groupHeader = win.locator('aside').getByText('Group Two').first();
+  if (await groupHeader.count()) {
+    const headerVisible = await groupHeader.isVisible().catch(() => false);
+    if (headerVisible) {
+      await groupHeader.click();
+      await win.waitForTimeout(120);
+      const stillTextarea = await activeIsTextarea();
+      const tag = await activeTag();
+      if (!stillTextarea && tag === 'BODY') {
+        throw new Error('clicking group header while typing dropped focus to <body>');
+      }
+      const value = await textarea.inputValue();
+      if (value !== 'drafting…') throw new Error(`group header click corrupted draft: got ${JSON.stringify(value)}`);
+    }
+  }
+
+  // Contract 4: open Settings → focus inside dialog; close → composer usable.
+  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
+  const settingsBtn = win.getByRole('button', { name: /^settings$/i }).first();
+  if (await settingsBtn.count()) {
+    await settingsBtn.click();
+    const dialog = win.locator('[role="dialog"]').first();
+    await dialog.waitFor({ state: 'visible', timeout: 3000 }).catch(() => { throw new Error('Settings dialog never opened'); });
+    const insideDialog = await win.evaluate(() => {
+      const d = document.querySelector('[role="dialog"]');
+      return !!(d && d.contains(document.activeElement));
+    });
+    if (!insideDialog) {
+      const tag = await activeTag();
+      throw new Error(`Settings open: focus not inside dialog, activeElement=${tag}`);
+    }
+    await win.keyboard.press('Escape');
+    await dialog.waitFor({ state: 'hidden', timeout: 3000 }).catch(() => { throw new Error('Settings dialog did not close on Esc'); });
+    await win.waitForTimeout(300);
+    await textarea.click();
+    await textarea.fill('post-modal');
+    const v = await textarea.inputValue();
+    if (v !== 'post-modal') throw new Error('composer not usable after closing Settings');
+    await textarea.fill('');
+  }
+
+  // Contract 5: focusInputNonce bump must NOT yank focus from another <input>.
+  await win.evaluate(() => {
+    const inp = document.createElement('input');
+    inp.id = '__probeInput';
+    inp.type = 'text';
+    document.body.appendChild(inp);
+    inp.focus();
+  });
+  const beforeBump = await win.evaluate(() => document.activeElement?.id);
+  if (beforeBump !== '__probeInput') throw new Error(`failed to focus probe input pre-bump (id=${beforeBump})`);
+  await win.evaluate(() => window.__ccsmStore.getState().bumpComposerFocus());
+  await win.waitForTimeout(120);
+  const afterBump = await win.evaluate(() => document.activeElement?.id);
+  if (afterBump !== '__probeInput') {
+    const tag = await win.evaluate(() => document.activeElement?.tagName);
+    throw new Error(`focusInputNonce bump stole focus: now id="${afterBump}" tag="${tag}"`);
+  }
+  await win.evaluate(() => document.getElementById('__probeInput')?.remove());
+
+  log('contracts 1-5: send / switch / group-click / modal / preserve-other-input');
+}
+
+// ---------- theme-toggle ----------
+// Theme dark↔light: class flip lands within a frame, body bg luminance
+// changes substantially, contrast remains readable. Restores theme to dark.
+async function caseThemeToggle({ win, log, registerDispose }) {
+  registerDispose(async () => {
+    // Restore to dark (the harness baseline) so subsequent cases aren't
+    // surprised by light-mode CSS variables.
+    await win.evaluate(() => {
+      try { window.__ccsmStore.getState().setTheme('dark'); } catch {}
+    });
+  });
+
+  async function snapshot() {
+    return await win.evaluate(() => {
+      const html = document.documentElement;
+      function parseLum(s) {
+        if (!s) return null;
+        let m = s.match(/^oklch\(\s*([0-9.]+)/i) || s.match(/^oklab\(\s*([0-9.]+)/i);
+        if (m) return parseFloat(m[1]);
+        m = s.match(/rgba?\(\s*(\d+)[, ]+(\d+)[, ]+(\d+)/);
+        if (m) {
+          const r = +m[1], g = +m[2], b = +m[3];
+          return (0.2126 * r + 0.7152 * g + 0.0722 * b) / 255;
+        }
+        return null;
+      }
+      const bgRaw = getComputedStyle(html).getPropertyValue('--color-bg-app').trim() ||
+        getComputedStyle(html).backgroundColor;
+      const fgRaw = getComputedStyle(html).getPropertyValue('--color-fg-primary').trim() ||
+        getComputedStyle(html).color;
+      const sidebar = document.querySelector('aside');
+      const sidebarBg = sidebar ? getComputedStyle(sidebar).backgroundColor : bgRaw;
+      const bgLum = parseLum(bgRaw);
+      const fgLum = parseLum(fgRaw);
+      return {
+        themeClassDark: html.classList.contains('dark'),
+        themeClassLight: html.classList.contains('theme-light'),
+        dataTheme: html.dataset.theme,
+        sidebarBg, fg: fgRaw,
+        contrast: bgLum != null && fgLum != null ? Math.abs(bgLum - fgLum) : 0,
+        bgLum: bgLum ?? -1
+      };
+    });
+  }
+
+  await win.evaluate(() => { window.__ccsmStore.getState().setTheme('dark'); });
+  await win.waitForTimeout(150);
+  const dark1 = await snapshot();
+  if (!dark1.themeClassDark || dark1.themeClassLight) throw new Error(`expected initial dark theme classes, got ${JSON.stringify(dark1)}`);
+  if (dark1.dataTheme !== 'dark') throw new Error(`html[data-theme] should be 'dark', got ${dark1.dataTheme}`);
+  if (dark1.contrast < 0.3) throw new Error(`dark-mode contrast too low (${dark1.contrast.toFixed(2)}). snapshot=${JSON.stringify(dark1)}`);
+
+  // Need a session for the sidebar Settings button to be visible? No — it
+  // renders regardless. But it is rendered as the first sidebar button.
+  const sidebarBtn = win.getByRole('button', { name: /^settings$/i }).first();
+  await sidebarBtn.click();
+  const dialog = win.getByRole('dialog');
+  await dialog.waitFor({ state: 'visible', timeout: 3000 });
+  const lightRadio = dialog.getByRole('radio', { name: /^light$/i });
+  await lightRadio.click();
+  await win.waitForFunction(
+    () => document.documentElement.classList.contains('theme-light'),
+    null,
+    { timeout: 1000 }
+  );
+  const light1 = await snapshot();
+  if (light1.themeClassDark) throw new Error('html.dark still set after switching to Light');
+  if (!light1.themeClassLight) throw new Error('html.theme-light not set after switching to Light');
+  if (light1.dataTheme !== 'light') throw new Error(`html[data-theme] should be 'light', got ${light1.dataTheme}`);
+  if (!(light1.bgLum > dark1.bgLum + 0.4)) {
+    throw new Error(`light-mode bg not noticeably brighter (dark=${dark1.bgLum.toFixed(2)}, light=${light1.bgLum.toFixed(2)})`);
+  }
+  if (light1.contrast < 0.3) throw new Error(`light-mode contrast too low (${light1.contrast.toFixed(2)})`);
+  if (light1.sidebarBg === 'rgba(0, 0, 0, 0)' || light1.sidebarBg === 'transparent') {
+    throw new Error(`sidebar background is transparent in light mode (${light1.sidebarBg})`);
+  }
+
+  const darkRadio = dialog.getByRole('radio', { name: /^dark$/i });
+  await darkRadio.click();
+  await win.waitForFunction(
+    () => document.documentElement.classList.contains('dark') && !document.documentElement.classList.contains('theme-light'),
+    null,
+    { timeout: 1000 }
+  );
+  const dark2 = await snapshot();
+  if (dark2.dataTheme !== 'dark') throw new Error(`html[data-theme] should be back to 'dark', got ${dark2.dataTheme}`);
+  if (!(dark2.bgLum < light1.bgLum - 0.4)) {
+    throw new Error(`dark-mode bg lum (${dark2.bgLum.toFixed(2)}) not noticeably darker than light (${light1.bgLum.toFixed(2)})`);
+  }
+
+  await win.keyboard.press('Escape');
+
+  log(`dark1=${dark1.bgLum.toFixed(2)} light=${light1.bgLum.toFixed(2)} dark2=${dark2.bgLum.toFixed(2)}`);
+}
+
+// ---------- language-toggle ----------
+// Live language flip via Settings → Appearance Language segmented; sidebar /
+// composer / settings strings flip between en and zh; protected English proper
+// nouns survive the translation. Restores language to en.
+async function caseLanguageToggle({ win, log, registerDispose }) {
+  registerDispose(async () => {
+    // Restore en so subsequent cases (which assert English UI) aren't broken.
+    // Language preference lives in localStorage under `ccsm:preferences` AND
+    // in i18next's runtime + the main-process mirror; we clear/reset all three
+    // to guarantee the next case (and the next harness launch) sees `en`.
+    await win.evaluate(async () => {
+      try { window.localStorage.removeItem('ccsm:preferences'); } catch {}
+      try { window.ccsm?.i18n?.setLanguage?.('en'); } catch {}
+      try {
+        const i18n = window.__ccsmI18n;
+        if (i18n && typeof i18n.changeLanguage === 'function') await i18n.changeLanguage('en');
+      } catch {}
+    });
+  });
+
+  const PROTECTED_TERMS = [
+    'MCP', 'CLI', 'IPC', 'API', 'URL', 'JSONL', 'JSON', 'SDK', 'REST',
+    'Claude', 'Anthropic', 'CCSM', 'Electron', 'GitHub'
+  ];
+
+  async function openSettingsAppearance() {
+    const dialog = win.getByRole('dialog');
+    if ((await dialog.count()) === 0) {
+      const btn = win.getByRole('button', { name: /^(settings|设置)$/i }).first();
+      await btn.waitFor({ state: 'visible', timeout: 5000 });
+      await btn.click();
+    }
+    await dialog.waitFor({ state: 'visible', timeout: 3000 });
+    const appearanceTab = dialog.getByRole('button', { name: /^(appearance|外观)$/i });
+    if (await appearanceTab.isVisible().catch(() => false)) await appearanceTab.click();
+    return dialog;
+  }
+  async function pickLanguage(dialog, name) {
+    const radio = dialog.getByRole('radio', { name });
+    await radio.click();
+  }
+  async function closeDialog() {
+    await win.keyboard.press('Escape');
+    await win.getByRole('dialog').waitFor({ state: 'hidden', timeout: 1500 }).catch(() => {});
+  }
+
+  // Force English baseline.
+  let dialog = await openSettingsAppearance();
+  await pickLanguage(dialog, /^english$/i);
+  await win.waitForTimeout(150);
+  await closeDialog();
+
+  async function snapshotStrings() {
+    return await win.evaluate(() => {
+      const settingsBtn = document.querySelector(
+        'aside button[aria-label="Settings"], aside button[aria-label="设置"]'
+      );
+      const settingsLabel =
+        (settingsBtn && (settingsBtn.getAttribute('aria-label') || settingsBtn.textContent || '').trim()) ||
+        Array.from(document.querySelectorAll('aside button')).map((b) => b.textContent?.trim() || '')
+          .find((t) => /Settings|设置/.test(t)) || null;
+      const newSessionBtn = document.querySelector(
+        'aside button[aria-label="New session"], aside button[aria-label="新会话"]'
+      );
+      const newSessionText =
+        (newSessionBtn && (newSessionBtn.getAttribute('aria-label') || newSessionBtn.textContent || '').trim()) ||
+        Array.from(document.querySelectorAll('aside button')).map((b) => b.textContent?.trim() || '')
+          .find((t) => /New Session|新会话/.test(t)) || null;
+      const ta = document.querySelector('textarea');
+      const placeholder = ta ? ta.getAttribute('placeholder') : null;
+      return { settingsLabel, newSessionText, placeholder };
+    });
+  }
+
+  const en1 = await snapshotStrings();
+  if (!en1.settingsLabel || !/Settings/i.test(en1.settingsLabel)) {
+    throw new Error(`English baseline: settings label not English. got: ${JSON.stringify(en1)}`);
+  }
+  if (en1.placeholder && /[一-鿿]/.test(en1.placeholder)) {
+    throw new Error(`English baseline: placeholder contains CJK chars. got: ${en1.placeholder}`);
+  }
+
+  dialog = await openSettingsAppearance();
+  await pickLanguage(dialog, /^中文$/);
+  await win.waitForTimeout(200);
+  await closeDialog();
+
+  const zh1 = await snapshotStrings();
+  if (!zh1.settingsLabel || !/设置/.test(zh1.settingsLabel)) {
+    throw new Error(`After zh switch: settings label not Chinese. got: ${JSON.stringify(zh1)}`);
+  }
+  if (zh1.newSessionText && !/[一-鿿]/.test(zh1.newSessionText)) {
+    throw new Error(`After zh switch: new session text contains no CJK. got: ${zh1.newSessionText}`);
+  }
+  if (zh1.placeholder && !/[一-鿿]/.test(zh1.placeholder)) {
+    throw new Error(`After zh switch: placeholder contains no CJK. got: ${zh1.placeholder}`);
+  }
+
+  dialog = await openSettingsAppearance();
+  await pickLanguage(dialog, /^english$/i);
+  await win.waitForTimeout(200);
+  await closeDialog();
+  const en2 = await snapshotStrings();
+  if (!en2.settingsLabel || !/Settings/i.test(en2.settingsLabel)) {
+    throw new Error(`After en switch back: settings label not English. got: ${JSON.stringify(en2)}`);
+  }
+
+  // Protected-terms parity scan.
+  const parity = await win.evaluate((terms) => {
+    const i18next = (window).__ccsmI18n;
+    if (!i18next || !i18next.store) return { error: 'i18next not exposed on window.__ccsmI18n' };
+    const enRes = i18next.store.data.en?.translation;
+    const zhRes = i18next.store.data.zh?.translation;
+    if (!enRes || !zhRes) return { error: 'translation namespace missing' };
+    const violations = [];
+    function walk(enNode, zhNode, prefix) {
+      if (typeof enNode === 'string') {
+        if (typeof zhNode !== 'string') return;
+        for (const term of terms) {
+          const re = new RegExp(`\\b${term}\\b`);
+          if (re.test(enNode)) {
+            if (!new RegExp(`\\b${term}\\b`).test(zhNode)) {
+              violations.push({ key: prefix, term, en: enNode, zh: zhNode });
+            }
+          }
+        }
+        return;
+      }
+      if (enNode && typeof enNode === 'object') {
+        for (const k of Object.keys(enNode)) {
+          walk(enNode[k], zhNode ? zhNode[k] : undefined, prefix ? `${prefix}.${k}` : k);
+        }
+      }
+    }
+    walk(enRes, zhRes, '');
+    return { violations };
+  }, PROTECTED_TERMS);
+
+  if (parity.error) throw new Error(`could not read i18n catalogs: ${parity.error}`);
+  if (parity.violations.length > 0) {
+    const sample = parity.violations.slice(0, 5).map((v) => `${v.key} [${v.term}] en="${v.en}" zh="${v.zh}"`).join('\n  ');
+    throw new Error(`${parity.violations.length} zh strings dropped a protected English proper noun:\n  ${sample}`);
+  }
+
+  log(`en→zh→en flip OK; protected-term parity 0 violations across ${PROTECTED_TERMS.length} terms`);
+}
+
+// ---------- i18n-settings-zh ----------
+// All four Settings panes (Appearance / Notifications / Updates / Connection)
+// render Chinese labels when language=zh. Restores language to en.
+async function caseI18nSettingsZh({ win, log, registerDispose }) {
+  registerDispose(async () => {
+    // Same restore strategy as language-toggle — strip the persisted
+    // preference so the next harness launch boots English again.
+    await win.evaluate(async () => {
+      try { window.localStorage.removeItem('ccsm:preferences'); } catch {}
+      try { window.ccsm?.i18n?.setLanguage?.('en'); } catch {}
+      try {
+        const i18n = window.__ccsmI18n;
+        if (i18n && typeof i18n.changeLanguage === 'function') await i18n.changeLanguage('en');
+      } catch {}
+    });
+  });
+
+  async function openSettings() {
+    const dialog = win.getByRole('dialog');
+    if ((await dialog.count()) === 0) {
+      const btn = win.getByRole('button', { name: /^(settings|设置)$/i }).first();
+      await btn.waitFor({ state: 'visible', timeout: 5000 });
+      await btn.click();
+    }
+    await dialog.waitFor({ state: 'visible', timeout: 3000 });
+    return dialog;
+  }
+  async function closeDialog() {
+    await win.keyboard.press('Escape');
+    await win.getByRole('dialog').waitFor({ state: 'hidden', timeout: 1500 }).catch(() => {});
+  }
+
+  let dialog = await openSettings();
+  const zhRadio = dialog.getByRole('radio', { name: /^中文$/ });
+  await zhRadio.waitFor({ state: 'visible', timeout: 3000 });
+  await zhRadio.click();
+  await win.waitForTimeout(250);
+
+  async function switchTab(name) {
+    const tab = dialog.getByRole('tab', { name });
+    await tab.waitFor({ state: 'visible', timeout: 2000 });
+    await tab.click();
+    await win.waitForTimeout(150);
+  }
+  function assertHasText(haystack, needle, where) {
+    if (!haystack.includes(needle)) {
+      throw new Error(`${where}: expected to find "${needle}" in pane text. Got snippet:\n${haystack.slice(0, 800)}`);
+    }
+  }
+  function assertNotHasText(haystack, needle, where) {
+    if (haystack.includes(needle)) {
+      throw new Error(`${where}: unexpected English string "${needle}" leaked into zh pane.\n${haystack.slice(0, 800)}`);
+    }
+  }
+  async function paneText() {
+    return await dialog.evaluate((el) => {
+      const main = el.querySelector('div.overflow-y-auto');
+      return ((main && main.textContent) || el.textContent || '').trim();
+    });
+  }
+
+  // Appearance.
+  let txt = await paneText();
+  assertHasText(txt, '主题', 'appearance');
+  assertHasText(txt, '字号', 'appearance');
+  assertHasText(txt, '密度', 'appearance');
+  assertNotHasText(txt, 'Theme', 'appearance');
+  assertNotHasText(txt, 'Density', 'appearance');
+
+  // Notifications.
+  await switchTab(/^通知$/);
+  txt = await paneText();
+  assertHasText(txt, '启用通知', 'notifications');
+  assertHasText(txt, '权限请求', 'notifications');
+  assertHasText(txt, '发送测试通知', 'notifications');
+  assertNotHasText(txt, 'Enable notifications', 'notifications');
+  assertNotHasText(txt, 'Test notification', 'notifications');
+
+  // Updates.
+  await switchTab(/^更新$/);
+  txt = await paneText();
+  assertHasText(txt, '版本', 'updates');
+  assertHasText(txt, '检查更新', 'updates');
+  assertHasText(txt, '自动检查', 'updates');
+  assertNotHasText(txt, 'Check for updates', 'updates');
+  assertNotHasText(txt, 'Automatic checks', 'updates');
+
+  // Connection.
+  await switchTab(/^连接$/);
+  txt = await paneText();
+  assertHasText(txt, '默认模型', 'connection');
+  assertHasText(txt, 'Auth Token', 'connection');
+  assertHasText(txt, '打开 settings.json', 'connection');
+  assertNotHasText(txt, 'Default model', 'connection');
+  assertNotHasText(txt, 'Open settings.json', 'connection');
+
+  await closeDialog();
+
+  log('Appearance / Notifications / Updates / Connection panes all render Chinese labels');
+}
+
 // ---------- harness spec ----------
 await runHarness({
   name: 'ui',
   setup: async ({ win }) => {
-    // Suppress the "Claude CLI not found" first-launch dialog.
-    await win.evaluate(() => {
-      window.__ccsmStore?.setState({
-      });
+    // Suppress the "Claude CLI not found" first-launch dialog AND force the
+    // i18n preference back to English on every harness boot. The language
+    // pref persists in localStorage (`ccsm:preferences`); a previous run
+    // that left zh persisted would otherwise break every English-anchored
+    // assertion in subsequent cases.
+    await win.evaluate(async () => {
+      try { window.localStorage.removeItem('ccsm:preferences'); } catch {}
+      try { window.ccsm?.i18n?.setLanguage?.('en'); } catch {}
+      try {
+        const i18n = window.__ccsmI18n;
+        if (i18n && typeof i18n.changeLanguage === 'function') await i18n.changeLanguage('en');
+      } catch {}
+      window.__ccsmStore?.setState({});
     });
   },
   cases: [
@@ -695,6 +1708,18 @@ await runHarness({
     { id: 'popover-cross-dismiss', run: casePopoverCrossDismiss },
     { id: 'type-scale-snapshot', run: caseTypeScaleSnapshot },
     { id: 'banner-i18n-toggle', run: caseBannerI18nToggle },
-    { id: 'toast-a11y', run: caseToastA11y }
+    { id: 'toast-a11y', run: caseToastA11y },
+    { id: 'cwd-popover-recent-unfiltered', run: caseCwdPopoverRecentUnfiltered },
+    { id: 'palette-empty', run: casePaletteEmpty },
+    { id: 'palette-nav', run: casePaletteNav },
+    { id: 'settings-open', run: caseSettingsOpen },
+    { id: 'search-shortcut-f', run: caseSearchShortcutF },
+    { id: 'tutorial', run: caseTutorial },
+    { id: 'titlebar', run: caseTitlebar },
+    { id: 'tray', run: caseTray },
+    { id: 'focus-orchestration', run: caseFocusOrchestration },
+    { id: 'theme-toggle', run: caseThemeToggle },
+    { id: 'language-toggle', run: caseLanguageToggle },
+    { id: 'i18n-settings-zh', run: caseI18nSettingsZh }
   ]
 });
