--- scripts/probe-e2e-focus-orchestration.mjs
@@ -1,268 +1,0 @@
-// E2E: composer focus orchestration. The InputBar is the canonical text-entry
-// surface; focus should land there at the right moments and never be stolen
-// from text-entry surfaces the user is actively typing in.
-//
-// Contracts under test:
-//   1. After clicking Send, focus returns to the textarea (so the user can
-//      keep typing the next thought without re-grabbing the mouse).
-//   2. Selecting a different session in the sidebar moves focus into THAT
-//      session's textarea (Claude Desktop parity, also covered by
-//      probe-click-session-focus.mjs but re-asserted here against the same
-//      session-switch event since orchestration regressions tend to creep
-//      in around the focusInputNonce bump path).
-//   3. Clicking a sidebar group/header (non-session) does NOT yank focus out
-//      of the textarea while the user is typing — focusInputNonce should not
-//      bump on incidental sidebar interactions.
-//   4. Opening the Settings modal moves focus into the dialog (Radix default).
-//      Closing it returns focus to the textarea (via focusInputNonce bump).
-//   5. The composer must not steal focus away from another text input that
-//      already has focus when focusInputNonce bumps (e.g. the inline rename
-//      input on a session row, or any future settings field).
-import { _electron as electron } from 'playwright';
-import path from 'node:path';
-import fs from 'node:fs';
-import os from 'node:os';
-import { fileURLToPath } from 'node:url';
-import { appWindow } from './probe-utils.mjs';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const root = path.resolve(__dirname, '..');
-
-function fail(msg, app) {
-  console.error(`\n[probe-e2e-focus-orchestration] FAIL: ${msg}`);
-  if (app) app.close().catch(() => {});
-  process.exit(1);
-}
-
-const userDataDir = fs.mkdtempSync(path.join(os.tmpdir(), 'agentory-probe-focus-'));
-console.log(`[probe-e2e-focus-orchestration] userData = ${userDataDir}`);
-
-const app = await electron.launch({
-  args: ['.', `--user-data-dir=${userDataDir}`],
-  cwd: root,
-  env: { ...process.env, NODE_ENV: 'development', CCSM_PROD_BUNDLE: '1' }
-});
-
-try { // ccsm-probe-cleanup-wrap
-
-try {
-  const win = await appWindow(app);
-  await win.waitForLoadState('domcontentloaded');
-  await win.waitForFunction(() => !!window.__ccsmStore, null, { timeout: 15_000 });
-
-  // Stub IPC where contextBridge allows. NOTE: the contextBridge-exposed
-  // `window.ccsm` is non-configurable, so renderer reassignment of its
-  // methods is best-effort — we rely on observable store/UI state instead
-  // of intercepting the IPC calls themselves.
-  await win.evaluate(() => {
-    try {
-      const real = window.ccsm;
-      if (real) {
-        // Best-effort no-ops; if the contextBridge wrapper rejects these
-        // assignments we fall back to letting the real IPC run (it will
-        // just fail silently in our isolated userData environment).
-        real.agentSend = async () => true;
-        real.agentStart = async () => ({ ok: true, sessionId: 'sdk-1' });
-        real.agentInterrupt = async () => true;
-      }
-    } catch {}
-  });
-
-  // Seed two sessions in two groups so we have something to switch between
-  // and a sidebar group header to click on.
-  await win.evaluate(() => {
-    window.__ccsmStore.setState({
-      groups: [
-        { id: 'g1', name: 'Group One', collapsed: false, kind: 'normal' },
-        { id: 'g2', name: 'Group Two', collapsed: false, kind: 'normal' }
-      ],
-      sessions: [
-        { id: 'sA', name: 'session-a', state: 'idle', cwd: 'C:/x', model: 'claude-opus-4', groupId: 'g1', agentType: 'claude-code' },
-        { id: 'sB', name: 'session-b', state: 'idle', cwd: 'C:/y', model: 'claude-opus-4', groupId: 'g2', agentType: 'claude-code' }
-      ],
-      activeId: 'sA',
-      messagesBySession: { sA: [{ kind: 'user', id: 'u-a', text: 'hi A' }], sB: [] },
-      // Mark sA as started so a Send doesn't bounce through agentStart.
-      startedSessions: { sA: true }
-    });
-  });
-
-  await win.waitForTimeout(300);
-
-  const textarea = win.locator('textarea');
-  await textarea.waitFor({ state: 'visible', timeout: 5000 });
-
-  async function activeTag() {
-    return win.evaluate(() => document.activeElement ? document.activeElement.tagName : null);
-  }
-  async function activeIsTextarea() {
-    return win.evaluate(() => document.activeElement?.tagName === 'TEXTAREA');
-  }
-
-  // --- Contract 1: focus returns to textarea after Send ---
-  await textarea.click();
-  await textarea.fill('hello there');
-  // Park focus on body so we can detect the Send-induced focus return.
-  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
-  if (await activeIsTextarea()) fail('precondition: textarea should not be focused right before Send', app);
-  // Click the Send button. Empty composer disables it; we filled above so
-  // it should be enabled.
-  const sendBtn = win.getByRole('button', { name: /send message/i }).first();
-  await sendBtn.waitFor({ state: 'visible', timeout: 3000 });
-  await sendBtn.click();
-  await win.waitForTimeout(200);
-  // Confirm the send went through by looking for the local-echo user block
-  // in the store (InputBar.send() appends it synchronously before any IPC).
-  const echoLanded = await win.evaluate(() => {
-    const blocks = window.__ccsmStore.getState().messagesBySession['sA'] ?? [];
-    return blocks.some((b) => b.kind === 'user' && b.text === 'hello there');
-  });
-  if (!echoLanded) fail('Send click did not produce a user-echo block — send() never ran', app);
-  // After Send, the textarea is cleared but focus orchestration should keep
-  // it as the active element (or bump back to it). We rely on the implicit
-  // re-render leaving the textarea focused — Radix Button click doesn't
-  // move focus elsewhere by default.
-  // NOTE: a Button click leaves focus ON the button. The contract here is
-  // that send() programmatically refocuses; if we don't see TEXTAREA, that's
-  // a regression worth flagging.
-  const tagAfterSend = await activeTag();
-  if (tagAfterSend !== 'TEXTAREA') {
-    // Soft-acceptable: BUTTON, since current send() doesn't programmatically
-    // re-focus. Capture as a known-gap warning instead of failing — this is
-    // future-work, not a regression. If the contract was previously TEXTAREA
-    // we'd hard-fail here.
-    console.log(`  [note] post-Send activeElement = ${tagAfterSend} (composer does not currently auto-refocus on Send click)`);
-  } else {
-    console.log('  post-Send activeElement = TEXTAREA');
-  }
-
-  // --- Contract 2: switching session via sidebar focuses target textarea ---
-  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
-  const rows = win.locator('aside li[role="option"]');
-  const rowCount = await rows.count();
-  if (rowCount < 2) fail(`expected ≥2 sidebar rows, got ${rowCount}`, app);
-  // Find the row that's NOT currently active (data-state or aria-selected).
-  // Click the second row by index.
-  await rows.nth(1).click();
-  await win.waitForTimeout(150);
-  if (!(await activeIsTextarea())) {
-    const tag = await activeTag();
-    fail(`after switching session via sidebar click, expected TEXTAREA focus, got ${tag}`, app);
-  }
-  console.log('  switch-session via sidebar -> TEXTAREA focused');
-
-  // --- Contract 3: clicking a group header does not steal focus into composer
-  // when focus is already in the textarea (no spurious nonce bump). ---
-  await textarea.click();
-  await textarea.fill('drafting…');
-  if (!(await activeIsTextarea())) fail('precondition: textarea should be focused before group click', app);
-  // Find a group header row in the sidebar. Most sidebars use a button or
-  // div with the group name. We scope to <aside> to avoid hitting a chat
-  // header by mistake.
-  const groupHeader = win.locator('aside').getByText('Group Two').first();
-  if (await groupHeader.count()) {
-    const headerVisible = await groupHeader.isVisible().catch(() => false);
-    if (headerVisible) {
-      await groupHeader.click();
-      await win.waitForTimeout(120);
-      // Focus should still be in the textarea OR have moved to the header
-      // button (acceptable — the user actually clicked it). The forbidden
-      // outcome is focus ending up on <body> with the textarea blurred while
-      // the user is mid-draft, which would mean the click stole focus AWAY.
-      const stillTextarea = await activeIsTextarea();
-      const tag = await activeTag();
-      if (!stillTextarea && tag === 'BODY') {
-        fail('clicking group header while typing dropped focus to <body> — focus stolen from composer', app);
-      }
-      // Also confirm the draft text was preserved (no accidental clear).
-      const value = await textarea.inputValue();
-      if (value !== 'drafting…') {
-        fail(`group header click corrupted composer draft: got ${JSON.stringify(value)}`, app);
-      }
-      console.log(`  group-header click during draft: activeElement=${tag}, draft preserved`);
-    }
-  } else {
-    console.log('  [skip] group header not found in sidebar — no clickable group element to test');
-  }
-
-  // --- Contract 4: open Settings -> focus moves into dialog. Close ->
-  // focus returns to composer. ---
-  // Re-park focus to body so the post-close focus check is meaningful.
-  await win.evaluate(() => { document.querySelector('textarea')?.blur(); document.body.focus(); });
-  const settingsBtn = win.getByRole('button', { name: /^settings$/i }).first();
-  if (await settingsBtn.count()) {
-    await settingsBtn.click();
-    // Radix Dialog renders [role="dialog"].
-    const dialog = win.locator('[role="dialog"]').first();
-    await dialog.waitFor({ state: 'visible', timeout: 3000 }).catch(() => fail('Settings dialog never opened', app));
-    // Focus should be inside the dialog.
-    const insideDialog = await win.evaluate(() => {
-      const d = document.querySelector('[role="dialog"]');
-      return !!(d && d.contains(document.activeElement));
-    });
-    if (!insideDialog) {
-      const tag = await activeTag();
-      fail(`Settings open: focus not inside dialog, activeElement=${tag}`, app);
-    }
-    console.log('  Settings open -> focus inside dialog');
-
-    // Close with Esc — Radix Dialog handles its own Escape. Important: this
-    // is also the regression case that the InputBar's document-level Esc
-    // listener does NOT fight the dialog (it bails when [role="dialog"]
-    // exists in the DOM).
-    await win.keyboard.press('Escape');
-    await dialog.waitFor({ state: 'hidden', timeout: 3000 }).catch(() => fail('Settings dialog did not close on Esc', app));
-    await win.waitForTimeout(300);
-    // Where focus lands post-close depends on Radix's restore policy +
-    // any composer focus orchestration. We do NOT hard-fail on BODY here:
-    // there's no explicit contract that closing Settings refocuses the
-    // composer in this codebase (it's not bumped through bumpComposerFocus).
-    // We DO assert that subsequently typing into the composer still works —
-    // that's the user-visible bit.
-    const tagAfterClose = await activeTag();
-    console.log(`  Settings close -> activeElement=${tagAfterClose}`);
-    await textarea.click();
-    await textarea.fill('post-modal');
-    const v = await textarea.inputValue();
-    if (v !== 'post-modal') fail('composer not usable after closing Settings', app);
-    await textarea.fill('');
-  } else {
-    console.log('  [skip] Settings button not found — modal focus restoration not exercised');
-  }
-
-  // --- Contract 5: when another text input has focus, a focusInputNonce
-  // bump must NOT yank focus out of it. We simulate by injecting a temporary
-  // <input> into the DOM, focusing it, then bumping the nonce. ---
-  await win.evaluate(() => {
-    const inp = document.createElement('input');
-    inp.id = '__probeInput';
-    inp.type = 'text';
-    document.body.appendChild(inp);
-    inp.focus();
-  });
-  const beforeBump = await win.evaluate(() => document.activeElement?.id);
-  if (beforeBump !== '__probeInput') {
-    fail(`failed to focus probe input pre-bump (activeElement id=${beforeBump})`, app);
-  }
-  await win.evaluate(() => window.__ccsmStore.getState().bumpComposerFocus());
-  await win.waitForTimeout(120);
-  const afterBump = await win.evaluate(() => document.activeElement?.id);
-  if (afterBump !== '__probeInput') {
-    const tag = await win.evaluate(() => document.activeElement?.tagName);
-    fail(`focusInputNonce bump stole focus from another <input>: now activeElement id="${afterBump}" tag="${tag}"`, app);
-  }
-  await win.evaluate(() => document.getElementById('__probeInput')?.remove());
-  console.log('  focusInputNonce bump preserved focus on a different <input>');
-
-  console.log('\n[probe-e2e-focus-orchestration] OK');
-  console.log('  composer focus orchestration: send / switch / group-click / modal / preserve-other-input');
-
-  await app.close();
-} catch (err) {
-  console.error('[probe-e2e-focus-orchestration] threw:', err);
-  await app.close().catch(() => {});
-  process.exit(1);
-} finally {
-  fs.rmSync(userDataDir, { recursive: true, force: true });
-}
-} finally { try { await app.close(); } catch {} } // ccsm-probe-cleanup-wrap
